scripts/One-2-3-45/scripts_in_container/requirements.txt
--find-links https://download.pytorch.org/whl/torch/
--find-links https://download.pytorch.org/whl/torchvision/
torch==2.0.1+cu118
torchvision==0.15.2+cu118
ninja
albumentations>=1.3.1
opencv-python>=4.8.0.76
pudb>=2022.1.3
imageio>=2.31.1
imageio-ffmpeg>=0.4.8
pytorch-lightning>=2.0.6
omegaconf>=2.3.0
test-tube>=0.7.5
streamlit>=1.25.0
einops>=0.6.1
torch-fidelity>=0.3.0
transformers>=4.31.0
kornia>=0.7.0
webdataset>=0.2.48
torchmetrics>=1.0.3
fire>=0.5.0
gradio==3.44.0
diffusers>=0.19.3
datasets[vision]>=2.14.4
rich>=13.5.2
plotly>=5.16.0
-e git+https://github.com/CompVis/taming-transformers.git#egg=taming-transformers
# elev est
dl_ext>=1.3.4
loguru>=0.7.0
matplotlib>=3.7.2
multipledispatch>=1.0.0
packaging>=23.1
Pillow>=9.3.0
PyYAML>=6.0.1
scikit_image>=0.21.0
scikit_learn>=1.3.0
scipy>=1.11.1
setuptools>=59.6.0
tensorboardX>=2.6.2
tqdm>=4.66.1
transforms3d>=0.4.1
trimesh>=3.23.1
yacs>=0.1.8
gdown>=4.7.1
git+https://github.com/NVlabs/nvdiffrast.git
git+https://github.com/openai/CLIP.git
# segment anything
onnxruntime>=1.15.1
onnx>=1.14.0
git+https://github.com/facebookresearch/segment-anything.git
# rembg
rembg>=2.0.50
# reconstruction
pyhocon>=0.3.60
icecream>=2.1.3
PyMCubes>=0.1.4
ninja>=1.11.1
# juypter
jupyter>=1.0.0
jupyterlab>=4.0.5
ipywidgets>=8.1.0
ipykernel>=6.25.1
panel>=1.2.1
jupyter_bokeh>=3.0.7
huggingface-hub==0.19.4
inplace-abn==1.0.12
git+https://github.com/mit-han-lab/torchsparse.git@v1.4.0
